% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{app-utils}
\alias{app-utils}
\alias{getAppData}
\title{App Utilities}
\usage{
getAppData(name = "employee_sales_kpis")
}
\arguments{
\item{name}{app name}
}
\description{
Functions used across all package apps
}
\section{Functions}{
\itemize{
\item \code{getAppData()}: retrieves an apps primary dataset given a valid app name

}}
\examples{
getAppData("employee_sales_kpis")

}
