% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlogData.R
\name{getBlogData}
\alias{getBlogData}
\title{Get App Dataset}
\usage{
getBlogData(name = "employee_sales_kpis")
}
\arguments{
\item{name}{app name}
}
\description{
Function used by all package apps to read the internally
stored appdata required by each. The primary app dataset
is returned given a valid name corresponding to an app
in this package
}
\examples{
getBlogData("employee_sales_kpis")

}
