% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{house_brands_kpis}
\alias{house_brands_kpis}
\title{Sales Performance of Retailer's House Brand}
\format{
## `house_brands_kpis`
A data table with 53,570 rows and 5 columns.
\describe{
  \item{store}{Anonymized Retail Location}
  \item{employee}{Anonymized Names of retailer employees}
  \item{period}{The time period associated with the KPI value}
  \item{kpi_name}{The name of the performance KPI metric}
  \item{kpi_value}{The calculated value of the performance KPI metric}
}
}
\source{
<Anonymized Proprietary Data>
}
\usage{
house_brands_kpis
}
\description{
Dataset is primary appdata for house_brands_kpis app included in this
package. This data contains a series of KPIs calculated across various time
periods. These performance metrics describe how well employees at a retail
location perform when selling the retailer's house branded products.
}
\keyword{datasets}
