% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{pkg-utils}
\alias{pkg-utils}
\alias{list_app_names}
\alias{list_app_deps}
\alias{has_app_deps}
\alias{is_app_valid}
\alias{get_app_dir}
\title{Utilities to Run Package Apps}
\usage{
list_app_names()

list_app_deps(name)

has_app_deps(name)

is_app_valid(name)

get_app_dir(name)
}
\arguments{
\item{name}{Name of app bundled in package}
}
\description{
Utilities to Run Package Apps
}
\section{Functions}{
\itemize{
\item \code{list_app_names()}: List the names of all apps in the package

\item \code{list_app_deps()}: List all dependencies for a package shiny app

\item \code{has_app_deps()}: Get list of app dependencies that are not found

\item \code{is_app_valid()}: Validates the name of package app

\item \code{get_app_dir()}: Get directory of package app

}}
\examples{
list_app_names()
list_app_deps("employee_sales_kpis")
}
