% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stockout_sales_impact}
\alias{stockout_sales_impact}
\title{Estimated Revenue Impact From Product Stockouts}
\format{
## `stockout_sales_impact`
A data table with 34,114 rows and 17 columns.
\describe{
  \item{org}{Anonymized Retailer Name}
  \item{store}{Anonymized Retail Location}
  \item{product_sku}{Product Sku}
  \item{order_date}{Date of Retail Order Sales}
  \item{category}{Product Category}
  \item{brand_name}{Product Brand}
  \item{tot_sales}{Total Sales over all time}
  \item{units_sold}{Total Unit Sales over all time}
  \item{has_sales}{Whether a product has sales on given order date}
  \item{period_units_sold}{Total units sold on order date}
  \item{period_tot_sales}{Total sales on order date}
  \item{ave_disc_r}{Average discount rate for product sales on order_date}
  \item{ave_ticket}{Average order size when product is included in order on given order_date}
  \item{wts}{Weights starting at 0 on day 1, ending at 1 on the last order_date for a given product}
  \item{tot_sales_est}{Estimated total sales on order_date assuming no product stockout}
  \item{c_sales_est}{Estimated cumulative total sales for days 1 through a given order_date assuming no product stockout}
  \item{c_sales_actual}{Actual cumulative total sales for days 1 through a given order_date with stockout events included}
}
}
\source{
<Anonymized Proprietary Data>
}
\usage{
stockout_sales_impact
}
\description{
Dataset is primary appdata for app stockout_sales_impact which is included
in this package. This data contains a timeseries of product sales statistics
including days in which a product was out of stock and not sold. Additionally,
dataset includes estimated total sales loss for days in which a stockout occurs.
}
\keyword{datasets}
