% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities}
\alias{utilities}
\alias{listBlogApps}
\alias{listBlogData}
\alias{read_desc_deps}
\alias{list_app_deps}
\alias{list_pkg_deps}
\alias{has_app_deps}
\alias{nf_package_deps}
\alias{stop_nf_depends}
\alias{is_app_valid}
\alias{get_app_dir}
\alias{is_ci}
\alias{is_testing}
\title{Utilities to Run Package Apps}
\usage{
listBlogApps()

listBlogData()

read_desc_deps(path)

list_app_deps(name)

list_pkg_deps()

has_app_deps(name)

nf_package_deps(name)

stop_nf_depends(nf_pkgs)

is_app_valid(name)

get_app_dir(name)

is_ci()

is_testing()
}
\arguments{
\item{path}{path to description file}

\item{name}{Name of app bundled in package}

\item{nf_pkgs}{vector of not found package dependencies}
}
\description{
Helper functions used to faciliate launching package apps or running
unit tests included in this package
}
\section{Functions}{
\itemize{
\item \code{listBlogApps()}: List the names of all apps in this package

\item \code{listBlogData()}: List the names of all app datasets in this package

\item \code{read_desc_deps()}: Reads dependencies declared in description files

\item \code{list_app_deps()}: List all dependencies for an app in this package

\item \code{list_pkg_deps()}: List all dependencies declared in this package

\item \code{has_app_deps()}: Show error if app dependencies not found

\item \code{nf_package_deps()}: Get list of app's R package depends that are not found

\item \code{stop_nf_depends()}: Show 'package dependencies not found' error

\item \code{is_app_valid()}: Show error if given name is not a valid package app

\item \code{get_app_dir()}: Get directory of an app included in this package

\item \code{is_ci()}: returns TRUE if called on CI

\item \code{is_testing()}: returns TRUE if called while unit testing

}}
\examples{
listBlogApps()

listBlogData()

}
