% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-run.R
\name{runBlogApp}
\alias{runBlogApp}
\title{Launch Package Shiny Apps}
\usage{
runBlogApp(name)
}
\arguments{
\item{name}{Name of app included in package. Lookup with \code{list_app_names()}}
}
\description{
Launch Package Shiny Apps
}
\examples{
\dontrun{
runBlogApp(name = "employee_sales_kpis")
}

}
