% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlogApp.R
\name{runBlogApp}
\alias{runBlogApp}
\title{Launch Package Apps}
\usage{
runBlogApp(name)
}
\arguments{
\item{name}{Name of app included in package. Lookup with \code{listBlogApps()}}
}
\description{
Launch Package Apps
}
\examples{
\dontrun{
runBlogApp(name = "employee_sales_kpis")
}

}
